requirements.txt
pandas==2.0.0
Pillow==9.5.0
APScheduler==3.10.1
PyQt6==6.5.0
